requirements.txt
aiohttp==3.7.3
discord.py==1.5.1
asyncpg==0.21.0
loguru==0.5.3
flake8==3.8.4
pytoml==0.1.21
aiofiles
